Bring next up to master (4.9.0). Next's Svelte 5 + TypeScript
architecture wins structurally: master's Vue components, composables,
pinia stores, and js/api modules landed as dead code and were removed.
Master's backend security hardening, pack-based offline translation
scaffolding, rrc search/bot-options additions, and helper
centralisation were merged into next's package layout.

Known follow-ups queued for port commits: Bergamot translation UI
(replaces old translator endpoints), relay bots/search pages, map tab
context menu, publish-site modal, archive card lazy rendering, and
lifecycle ports of master's inline meshchat.py improvements.

+++ meshchatx/src/backend/interface_editor.py
@@ -92,13 +92,13 @@ def coerce_rnode_frequency_hz(value: Any) -> Any:
             break
     hz = float(text) * scale
     if hz <= 0:
-        return int(round(hz))
+        return round(hz)
     if hz >= 1_000_000:
-        return int(round(hz))
+        return round(hz)
     whole = abs(hz - round(hz)) < 1e-9
     if (not whole) or (whole and hz < 10_000):
         hz *= 1_000_000.0
-    return int(round(hz))
+    return round(hz)
 
 
 def normalize_rnode_txpower(value: Any) -> Any:
